pyproject.toml
[project]
name = "ai_git_utils"
version = "0.5.0"
description = "AI git utils is an intelligent Git commit assistant that leverages AI to enhance your Git workflow."
keywords = ["git", "ai", "aigit"]
authors = [
    { name = "curry tang", email = "twn39@163.com" }
]
maintainers = [
    { name = "curry tang", email = "twn39@163.com" }
]

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.21.0",
    "gitpython>=3.1.46",
    "rich>=14.2.0",
    "openai>=2.8.1",
    "pip>=25.3",
]

readme = "README.md"
requires-python = ">= 3.11"

[project.urls]
"Homepage" = "https://github.com/twn39/aigit"
"Documentation" = "https://github.com/twn39/aigit"
"Repository" = "https://github.com/twn39/aigit"
"Source Code" = "https://github.com/twn39/aigit"

[project.scripts]
aigit = "ai_git_utils.main:app"

[build-system]
requires = ["hatchling==1.26.3"]
build-backend = "hatchling.build"


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["ai_git_utils"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.8",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "responses>=0.25.0",
    "freezegun>=1.5.0",
    "python-dotenv>=1.0.0",
]
